+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-28T04:23:34.300Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-28T17:03:47.548Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")