+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG1DFQHA9G5RMGX9B5VJJJY3"
+        "01KG2RZQXWVHXSANH80W3TQW09"
     )
     r"""The unique identifier of the data source"""