+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -134,7 +134,7 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-28T04:23:37.561Z")
+    created: Optional[datetime] = parse_datetime("2026-01-28T17:03:50.851Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")