+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG1DFQH9PXQ2KVWY2JSRHJDB"
+        "01KG2RZQXV7AW8K26RHSVSGPJ2"
     )
     r"""The unique identifier of the data source"""