+++ packages/orq-rc/RELEASES.md
@@ -5079,3 +5079,13 @@ Based on:
 - [python v4.3.0-rc.6] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.6 - packages/orq-rc
+
+## 2026-01-28 17:06:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.7] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.7 - packages/orq-rc